inSPIRE-1.3 pylint and code review removal of prints and temporary csv creation
